tests/wpt/mozilla/tests/bluetooth/getCharacteristics/characteristics-not-found.https.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/_mozilla/bluetooth/bluetooth-helpers.js"></script>
<script>
'use strict';
promise_test(t => {
    window.testRunner.setBluetoothMockDataSet(adapter_type.missing_characteristic_heart_rate);
    return window.navigator.bluetooth.requestDevice({
        filters: [{services: [heart_rate.name]}]
    })
    .then(device => device.gatt.connect())
    .then(gattServer => gattServer.getPrimaryService(heart_rate.name))
    .then(service => promise_rejects_dom(t, 'NotFoundError', service.getCharacteristics()));
}, 'Request for absent characteristics. Reject with NotFoundError.');
</script>
